[TAO-2501][Fix] Drop the spec generator; the template is the only source of defaults
`prepare_kpi_analyze_spec.py` loaded the template and then overwrote its `kpi`
block from argparse defaults unconditionally, so the template's values for those
fields could never take effect — including the skill's own. Three of them
disagreed with it:

    field               template   argparse
    conf_threshold      0.0        0.3
    num_recall_points   101        11
    ignore_sqwidth      0          40

A user running the generator with no tuning flags got a spec contradicting both
the template and the guidance in SKILL.md beside it.

Removed rather than patched: two places holding a default is the defect, and one
of them has to go. `validate_config` and `load_yaml` move into
`verify_kpi_analyze_spec.py`, which imported them from the deleted module, so
validation now runs on the artifact that will actually execute regardless of how
it was produced — previously a hand-written spec and a generated one took
different checks.

Filling the template is now the documented path. `verify` also reports
conf_threshold, num_recall_points and ignore_sqwidth, so the settings behind a
result are recoverable from its output.
